Pixel classifier can be trained from multiple images
Better separation of UI and processing code for pixel classification; processing moved to the processing module.
